Update jira package structure
