+++ qtaf-xray-plugin/src/main/java/de/qytera/qtaf/xray/builder/XrayJsonImportBuilder.java
@@ -12,9 +12,9 @@ import de.qytera.qtaf.xray.builder.test.MultipleIterationsXrayTestEntityBuilder;
 import de.qytera.qtaf.xray.builder.test.SingleIterationXrayTestEntityBuilder;
 import de.qytera.qtaf.xray.builder.test.XrayTestEntityBuilder;
 import de.qytera.qtaf.xray.config.XrayConfigHelper;
-import de.qytera.qtaf.xray.dto.request.issues.AdditionalField;
+import de.qytera.qtaf.xray.dto.request.jira.issues.AdditionalField;
 import de.qytera.qtaf.xray.dto.request.xray.ImportExecutionResultsRequestDto;
-import de.qytera.qtaf.xray.dto.response.issues.JiraIssueResponseDto;
+import de.qytera.qtaf.xray.dto.response.jira.issues.JiraIssueResponseDto;
 import de.qytera.qtaf.xray.entity.XrayTestEntity;
 import de.qytera.qtaf.xray.entity.XrayTestExecutionInfoEntity;
 import de.qytera.qtaf.xray.repository.jira.JiraIssueRepository;
