+++ qtaf-xray-plugin/src/main/java/de/qytera/qtaf/xray/repository/jira/JiraIssueRepository.java
@@ -6,10 +6,10 @@ import de.qytera.qtaf.core.gson.GsonFactory;
 import de.qytera.qtaf.http.RequestBuilder;
 import de.qytera.qtaf.http.WebService;
 import de.qytera.qtaf.xray.config.XrayConfigHelper;
-import de.qytera.qtaf.xray.dto.request.issues.AdditionalField;
-import de.qytera.qtaf.xray.dto.request.issues.JiraIssueSearchRequestDto;
-import de.qytera.qtaf.xray.dto.response.issues.JiraIssueResponseDto;
-import de.qytera.qtaf.xray.dto.response.issues.JiraIssueSearchResponseDto;
+import de.qytera.qtaf.xray.dto.request.jira.issues.AdditionalField;
+import de.qytera.qtaf.xray.dto.request.jira.issues.JiraIssueSearchRequestDto;
+import de.qytera.qtaf.xray.dto.response.jira.issues.JiraIssueResponseDto;
+import de.qytera.qtaf.xray.dto.response.jira.issues.JiraIssueSearchResponseDto;
 import jakarta.ws.rs.core.HttpHeaders;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
